+++ elmyra.ip.access.epo/elmyra/ip/util/render/phantomjs.py
@@ -14,7 +14,7 @@ rasterize_js = resource_filename(__name__, 'phantomjs_rasterize.js')
 def render_pdf(url):
     tmpfile = NamedTemporaryFile(suffix='.pdf')
     pdf = tmpfile.name
-    command = "phantomjs {rasterize_js} '{url}' {pdf} A4".format(rasterize_js=rasterize_js, url=url, pdf=pdf)
+    command = "phantomjs --ignore-ssl-errors=true {rasterize_js} '{url}' {pdf} A4".format(rasterize_js=rasterize_js, url=url, pdf=pdf)
     log.info('Rendering PDF: ' + command)
     subprocess.check_call(shlex.split(command))
     tmpfile.seek(0)